requirements.txt
aioredis==1.3.1
aniso8601==7.0.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
autobahn==21.11.1
Automat==20.2.0
certifi==2021.10.8
cffi==1.15.0
channels==3.0.4
channels-redis==3.3.1
charset-normalizer==2.0.7
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==35.0.0
daphne==3.0.2
defusedxml==0.7.1
Django==3.2.9
django-allauth==0.46.0
django-cors-headers==3.10.1
django-filter==21.1
django-rest-auth==0.9.5
django-safedelete==1.1.1
django-widget-tweaks==1.4.9
djangorestframework==3.12.4
drf-yasg==1.20.0
graphene==2.1.9
graphene-django==2.15.0
graphql-core==2.3.2
graphql-relay==2.0.1
hiredis==2.0.0
hyperlink==21.0.0
idna==3.3
incremental==21.3.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.0.3
MarkupSafe==2.0.1
msgpack==1.0.3
oauthlib==3.1.1
packaging==21.3
Pillow==8.4.0
promise==2.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
PyJWT==2.3.0
pyOpenSSL==21.0.0
pyparsing==3.0.6
python3-openid==3.2.0
pytz==2021.3
requests==2.26.0
requests-oauthlib==1.3.0
ruamel.yaml==0.17.17
ruamel.yaml.clib==0.2.6
Rx==1.6.1
service-identity==21.1.0
singledispatch==3.7.0
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
Twisted==21.7.0
txaio==21.2.1
typing-extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.7
zope.interface==5.4.0
